+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Extensions
 - `Set.insert(contentsOf:)` method is added, that inserts `Sequence` in the `Set`
 - Setter is added to `StringProtocol.subscript(_:)`
 
+Global Functions
+
+- `FIXME(_:file:line:function:)` method is added
+
 ### [5.2.1(88)](https://github.com/VakhoKontridze/VCore/releases/tag/5.2.1) — *2023 12 08*
 
 Other